+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Add `attach_ticket_label`, `detach_ticket_label`, `delete_ticket`, `set_custom_data`, `delete_message`,
   `send_ticket_message`, `send_ticket_media_message` (untested), `get_messages`, `get_message`,
   `store_custom_channel_message` (untested)
+* Add type hints on `get_webhook`, `get_help_center`, `get_help_center_category`, `get_help_center_article`,
+  `get_help_center_block`
 
 ## 0.1.2 (2024/09/12)